requirements.txt
pygraphml
Pillow
opencv-python
scipy
numpy